Add company size information
When the companies I work for largely aren't known, the names don't
really tell a lot about the context I was working in.  I've been
thinking about adding size, financial status, etc. to help convey this;
this is a first stab at that.  It will require a bit more thinking on
how to portray this in a way I like.